chore(client-app): refactor folder structure
