+++ client-app/__tests__/app/layout/App.spec.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import App from '../src/App';
+import App from '../../../src/app/layout/App';
 
 describe('App', () => {
   test('renders Reactivities heading', () => {
