+++ client-app/src/app/layout/App.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useState } from 'react';
-import './App.css';
 import axios from 'axios';
 import { Header, List } from 'semantic-ui-react';
