+++ client-app/src/main.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App.tsx';
+import App from './app/layout/App.tsx';
 import 'semantic-ui-css/semantic.min.css';
-import './index.css';
+import './app/layout/styles.css';
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
